// src/types/movies.type.ts
export interface Movies {
  serial: string
  title: string
  summary: string
  directors: string[]
  screenwriters: string[]
  producers: string[]
  cinematographers: string[]
  editors: string[]
  distributors: string[]
  music_composers: string[]
  release_date: string
  running_time: string
  budget: string
  box_office: string
  rating: string
  trailer: string
  poster: string
  wiki: string
}